Serialize works
